update link path

+++ ressources/php/erase_expire_token.php
@@ -10,15 +10,17 @@ $expireStatement = $db->prepare($query);
 $expireStatement->execute();
 $expires = $expireStatement->fetchAll(PDO::FETCH_ASSOC);
 
-for ($i = 0; $i < count($expires); ++$i) {
-    $str_expires = $expires[$i]['expires'];
-    $tmp = strtotime($str_expires);
-    if ($tmp < $now) {
-        $query = "DELETE FROM auth_tokens WHERE expires = :expires";
-        $rowStatement = $db->prepare($query);
-        $rowStatement->execute([
-            'expires' => $str_expires,
-        ]);
+if ($expires != false){
+    for ($i = 0; $i < count($expires); ++$i) {
+        $str_expires = $expires[$i]['expires'];
+        $tmp = strtotime($str_expires);
+        if ($tmp < $now) {
+            $query = "DELETE FROM auth_tokens WHERE expires = :expires";
+            $rowStatement = $db->prepare($query);
+            $rowStatement->execute([
+                'expires' => $str_expires,
+            ]);
+        }
     }
 }
 ?>
